+++ src/test/kotlin/com/dishan/raytracer/rays/CameraTest.kt
@@ -14,7 +14,7 @@ class CameraTest {
     fun `Constructing a camera`() {
         val hSize = 160
         val vSize = 120
-        val fieldOfView = (PI / 2).toFloat()
+        val fieldOfView = PI / 2
 
         val c = camera(hSize, vSize, fieldOfView)
 
@@ -26,21 +26,21 @@ class CameraTest {
 
     @Test
     fun `The pixel size for a horizontal canvas`() {
-        val c = camera(200, 125, (PI / 2).toFloat())
+        val c = camera(200, 125, PI / 2)
 
-        assert(c.pixelSize `~==` 0.01f)
+        assert(c.pixelSize `~==` 0.01)
     }
 
     @Test
     fun `The pixel size for a vertical canvas`() {
-        val c = camera(125, 200, (PI / 2).toFloat())
+        val c = camera(125, 200, PI / 2)
 
-        assert(c.pixelSize `~==` 0.01f)
+        assert(c.pixelSize `~==` 0.01)
     }
 
     @Test
     fun `Constructing a ray through the center of the canvas`() {
-        val c = camera(201, 101, (PI / 2).toFloat())
+        val c = camera(201, 101, PI / 2)
         val r = c.rayForPixel(100, 50)
 
         assert(r.origin `~==` point(0, 0, 0))
@@ -49,11 +49,11 @@ class CameraTest {
 
     @Test
     fun `Constructing a ray through a corner of the canvas`() {
-        val c = camera(201, 101, (PI / 2).toFloat())
+        val c = camera(201, 101, PI / 2)
         val r = c.rayForPixel(0, 0)
 
         assert(r.origin `~==` point(0, 0, 0))
-        assert(r.direction `~==` vector(0.66519f, 0.33259f, -0.66851f))
+        assert(r.direction `~==` vector(0.66519, 0.33259, -0.66851))
     }
 
     @Test
@@ -61,14 +61,14 @@ class CameraTest {
         val c = camera(
             hSize = 201,
             vSize = 101,
-            fieldOfView = (PI / 2).toFloat(),
+            fieldOfView = PI / 2,
             transform = rotationY(PI / 4) * translation(0, -2, 5),
         )
 
         val r = c.rayForPixel(100, 50)
 
         assert(r.origin `~==` point(0, 2, -5))
-        assert(r.direction `~==` vector(sqrt(2f) / 2, 0.0f, -sqrt(2f) / 2))
+        assert(r.direction `~==` vector(sqrt(2.0) / 2, 0.0, -sqrt(2.0) / 2))
     }
 
     @Test
@@ -80,9 +80,9 @@ class CameraTest {
         val up = vector(0, 1, 0)
         val transform = viewTransform(from, to, up)
 
-        val c = camera(11, 11, (PI / 2).toFloat(), transform)
+        val c = camera(11, 11, PI / 2, transform)
         val image = c.render(w)
 
-        assert(image[5, 5] `~==` Color(0.38066f, 0.47583f, 0.2855f))
+        assert(image[5, 5] `~==` Color(0.38066, 0.47583, 0.2855))
     }
 }
